+++ docs/whatsnew.md
@@ -1,2 +1,6 @@
+ #### 1.0.0
+ * Change translator behaviour when provided with non-string path or when path is not provided
+ * BREAKING CHANGE - helper methods are now chainable (refer to the readme to learn how to use them)
+ 
  #### 0.1.4
- * Fixed issues with undefined `this` value 
+ * Fixed issues with undefined `this` value